Make OperationRef serializable
