+++ src/diagram.rs
@@ -20,6 +20,7 @@ mod fork_clone_schema;
 mod fork_result_schema;
 mod join_schema;
 mod node_schema;
+mod operation_ref;
 mod registration;
 mod scope_schema;
 mod section_schema;
@@ -39,6 +40,7 @@ use fork_result_schema::{DynForkResult, ForkResultSchema};
 pub use join_schema::JoinOutput;
 use join_schema::{JoinSchema, SerializedJoinSchema};
 pub use node_schema::NodeSchema;
+pub use operation_ref::*;
 pub use registration::*;
 pub use scope_schema::*;
 pub use section_schema::*;
